requirements.txt
et-xmlfile==1.1.0
lxml==4.9.2
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==0.8.11
typing_extensions==4.6.3
